requirements.txt
pandas==2.2.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
openai==1.68.2
pyarrow==15.0.0
langchain==0.3.22
langchain-openai==0.3.11
langchain-community==0.3.20
langchain-core==0.3.49